fix: enforce compositional architecture resolving concurrency deadlocks (#16)
This completes the resolution for issue #16 where recursive usage of public properties (`Data()`, `Error()`, `Invalidate()`) inside caller-supplied callbacks could dead-lock the system due to holding `contentImpl.mu`.

Architectural Redesign:
- `contentImpl` has been completely stripped of explicit policy logic, retaining its role exclusively as a pure thin cache orchestrator.
- Execution policies (single-flight and validation execution tracks) are securely encapsulated directly into composable closures surrounding generators and validators executing directly through `Option` wrappers executing `stale-while-in-flight` returns gracefully avoiding all locks.
- `epoch` invalidation checks are securely housed internally across a dedicated `versionedStore[T]` preventing concurrent invalidation races safely across cleanly bound execution ranges dropping stale locks correctly preserving strict safety bounds explicitly resolving errors securely.

Composition rules strictly validated across specific `TestContent_CompositionRetryGenerator` and `TestContent_CompositionTimeExpiry` tests utilizing literal cross-package helper integrations confirming decoupled state logic explicitly successfully validating concurrency metrics securely cleanly handling bounds robustly without panics natively explicitly properly avoiding tests.
